Adding a 'unless file already exists' check

+++ install.sh
@@ -41,7 +41,7 @@ then
 fi
 
 # Jinja post-download installation
-if [[ -d ~/.vim/bundle/Jinja ]]
+if [[ -d ~/.vim/bundle/Jinja ]] && [[ ! -f ~/.vim/bundle/Jinja/syntax/htmljinja.vim ]]
 then
     curl http://www.vim.org/scripts/download_script.php?src_id=6961 -o ~/.vim/bundle/Jinja/syntax/htmljinja.vim
 fi
